fix(lint): Replace eslint-disable comments with proper boundary rules
Move AccountRole enum and AccessType constants to @/infra/auth/ layer.

Update ESLint config with targeted exemptions for legitimate

cross-layer patterns (server components, route handlers, LLM,

config bootstrapping). Remove all 15 eslint-disable comments.
