+++ src/app/(frontend)/exercises/[id]/page.tsx
@@ -1,6 +1,5 @@
 import React from 'react'
 
-// eslint-disable-next-line no-restricted-imports -- TODO: migrate to proper layer
 import { getPayload } from 'payload'
 import configPromise from '@payload-config'
 import { notFound } from 'next/navigation'
