+++ src/app/(frontend)/layout.tsx
@@ -22,7 +22,6 @@ import { defaultLocale, getDirection } from '@/i18n/config'
 import { getSystemLocale } from '@/i18n/server-locale'
 import { I18nProvider } from '@/ui/web/providers/I18n'
 
-// eslint-disable-next-line no-restricted-imports -- TODO: migrate to proper layer
 import { getPayload } from 'payload'
 import config from '@payload-config'
 import './globals.css'
