+++ src/app/api/blob/upload-token/route.ts
@@ -15,7 +15,6 @@ import {
 } from '@/server/chat-assets/constants'
 import { buildChatAssetPathname } from '@/server/chat-assets/pathname'
 
-// eslint-disable-next-line no-restricted-imports -- TODO: migrate to proper layer
 import { getDefaultTenantId } from '@/server/repos/tenant/get-default-tenant'
 
 const clientPayloadSchema = z.object({
