+++ src/app/api/chapters/by-grade/route.ts
@@ -1,7 +1,6 @@
 import '@/infra/config/server-init'
 
 import { NextRequest, NextResponse } from 'next/server'
-// eslint-disable-next-line no-restricted-imports -- TODO: migrate to service layer
 import { queryChaptersByGrade } from '@/server/repos/queries/chapters'
 import { getPayload } from 'payload'
 import configPromise from '@payload-config'
