+++ src/infra/llm/providers/factory.ts
@@ -16,7 +16,6 @@ import {
 } from '@/infra/config/runtime/config-values'
 import { logger } from '@/infra/utils/logger'
 
-// eslint-disable-next-line no-restricted-imports -- TODO: migrate to proper layer
 import { getDefaultTenantId } from '@/server/repos/tenant/get-default-tenant'
 import type { Payload } from 'payload'
 import { LLMProviderType } from './types'
