+++ src/infra/llm/providers/shared/media-reader.ts
@@ -12,7 +12,6 @@ import { isVercelBlobUrl } from '@/infra/blob/vercel-blob-adapter'
 import type { MediaPartWithPath } from '@/infra/llm/multimodal/types'
 import { logger } from '@/infra/utils/logger'
 
-// eslint-disable-next-line no-restricted-imports -- TODO: migrate to proper layer
 import { normalizeToAbsoluteUrl } from '@/server/services/pdf-fetcher'
 
 export interface MediaReaderOptions {
